requirements.txt
# Core dependencies
numpy>=1.24.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0

# LLM providers (install the one you need)
openai>=1.0.0  # For OpenAI GPT models
anthropic>=0.18.0  # For Anthropic Claude models

# Data handling and analysis
pandas>=2.0.0  # For data manipulation
matplotlib>=3.7.0  # For visualization
seaborn>=0.12.0  # For statistical plots
scipy>=1.10.0  # For statistical analysis

# Development tools
pytest>=7.4.0  # For testing
pytest-cov>=4.1.0  # For test coverage
black>=23.0.0  # For code formatting
flake8>=6.0.0  # For linting
mypy>=1.4.0  # For type checking

# Documentation
sphinx>=7.0.0  # For documentation generation
sphinx-rtd-theme>=1.3.0  # Documentation theme

# Optional dependencies for advanced features
# networkx>=3.0  # For network analysis of agent interactions
# scikit-learn>=1.3.0  # For clustering and ML analysis
# plotly>=5.15.0  # For interactive visualizations